Move test files to match package split
ExtremeConditionDialogTest and ReattachControllerTest were left in
the canvas root package after the sub-package split, causing test
compilation failures due to missing imports and visibility.
